+++ weaver/wps_restapi/swagger_definitions.py
@@ -6459,6 +6459,25 @@ class CWLIdentifier(ProcessIdentifier):
     )
 
 
+class CWLGraphEntryPoint(ExtendedSchemaNode):
+    schema_type = String
+    description = "CWL $graph entry point identifier with # prefix (e.g., '#main')."
+    example = "#main"
+    pattern = r"^#[A-Za-z0-9]+(?:[-_][A-Za-z0-9]+)*$"
+
+
+class CWLGraphIdentifier(AnyOfKeywordSchema):
+    """
+    Identifier for a CWL item within a $graph.
+    Can be a #-prefixed entry point, a UUID, or a standard process ID.
+    """
+    _any_of = [
+        CWLGraphEntryPoint(),
+        UUID(),
+        SLUG(),
+    ]
+
+
 class CWLIntentURL(URL):
     description = (
         "Identifier URL to a concept for the type of computational operation accomplished by this Process "
@@ -7271,7 +7290,12 @@ class ExecutionUnit(OneOfKeywordSchema):
 
 class ExecutionUnitList(ExtendedSequenceSchema):
     item = ExecutionUnit(name="ExecutionUnit")
-    validator = Length(min=1, max=1)
+    validator = Length(min=1)
+    description = (
+        "List of execution units to deploy. "
+        "When multiple units are provided, they should represent workflow steps and/or the main workflow. "
+        "Deployment order will be determined automatically based on dependencies."
+    )
 
 
 class ProcessDeploymentWithContext(ProcessDeployment):
@@ -7378,26 +7402,22 @@ class DeployContextDefinition(NotKeywordSchema, DeployParameters):
 
 
 class CWLGraphItem(CWLApp):  # no 'cwlVersion', only one at the top
-    id = CWLIdentifier()  # required in this case
+    id = CWLGraphIdentifier()  # required in this case
 
 
 class CWLGraphList(ExtendedSequenceSchema):
     cwl = CWLGraphItem()
 
 
-# FIXME: supported nested and $graph multi-deployment (https://github.com/crim-ca/weaver/issues/56)
 class CWLGraphBase(ExtendedMappingSchema):
     graph = CWLGraphList(
         name="$graph", description=(
-            "Graph definition that defines *exactly one* CWL Application Package represented as list. "
-            "Multiple definitions simultaneously deployed is NOT supported currently."
-            # "Graph definition that combines one or many CWL Application Packages within a single payload. "
-            # "If a single application is given (list of one item), it will be deployed as normal CWL by itself. "
-            # "If multiple applications are defined, the first MUST be the top-most Workflow process. "
-            # "Deployment of other items will be performed, and the full deployment will be persisted only if all are "
-            # "valid. The resulting Workflow will be registered as a package by itself (i.e: not as a graph)."
+            "Graph definition that combines one or many CWL Application Packages within a single payload. "
+            "If a single application is given (list of one item), it will be deployed as a single process. "
+            "If multiple applications are defined, at least one must be a Workflow. The order does not matter "
+            "as child processes (CommandLineTool) are automatically deployed first, then the parent Workflow."
         ),
-        validator=Length(min=1, max=1)
+        validator=Length(min=1)
     )